refactor: implement orchestra/server/runtime/manager separation (Phase 1-4)
- Create runtime/ module for event loop and execution machinery
- Move dispatcher/ and orchestra/ core to runtime/
- Update imports and remove old shims
- Reorganize tests to align with new module structure
