--- src/vibe3/orchestra/serve.py
@@ -1,6 +1,0 @@
-"""Shim for vibe3.orchestra.serve - moved to vibe3.server.app/registry."""
-
-from vibe3.server.app import _run, app, start, status, stop
-from vibe3.server.registry import _build_server
-
-__all__ = ["app", "start", "stop", "status", "_run", "_build_server"]
