+++ src/vibe3/orchestra/services/assignee_dispatch.py
@@ -13,8 +13,8 @@ from vibe3.manager.manager_executor import ManagerExecutor
 from vibe3.models.orchestration import IssueInfo
 from vibe3.orchestra.config import OrchestraConfig
 from vibe3.orchestra.dependency_checker import DependencyChecker
-from vibe3.orchestra.event_bus import GitHubEvent, ServiceBase
 from vibe3.orchestra.services.status_service import OrchestraStatusService
+from vibe3.runtime.event_bus import GitHubEvent, ServiceBase
 
 _PRIORITY_MAP: dict[str, int] = {
     "priority/urgent": 0,
