+++ src/vibe3/orchestra/services/comment_reply.py
@@ -8,7 +8,7 @@ from loguru import logger
 
 from vibe3.clients.github_client import GitHubClient
 from vibe3.orchestra.config import OrchestraConfig
-from vibe3.orchestra.event_bus import GitHubEvent, ServiceBase
+from vibe3.runtime.event_bus import GitHubEvent, ServiceBase
 
 
 def _build_mention_pattern(usernames: list[str]) -> re.Pattern[str]:
