+++ src/vibe3/orchestra/services/pr_review_dispatch.py
@@ -10,7 +10,7 @@ from loguru import logger
 
 from vibe3.manager.manager_executor import ManagerExecutor
 from vibe3.orchestra.config import OrchestraConfig
-from vibe3.orchestra.event_bus import GitHubEvent, ServiceBase
+from vibe3.runtime.event_bus import GitHubEvent, ServiceBase
 
 
 class PRReviewDispatchService(ServiceBase):
