+++ src/vibe3/runtime/__init__.py
@@ -1,0 +1,4 @@
+"""Runtime module for Vibe Orchestra.
+
+Contains event loop, execution engine, and circuit breaker.
+"""
