+++ src/vibe3/runtime/circuit_breaker.py
