+++ src/vibe3/runtime/heartbeat.py
@@ -6,7 +6,7 @@ import os
 from loguru import logger
 
 from vibe3.orchestra.config import OrchestraConfig
-from vibe3.orchestra.event_bus import GitHubEvent, ServiceBase
+from vibe3.runtime.event_bus import GitHubEvent, ServiceBase
 
 
 class HeartbeatServer:
