+++ tests/vibe3/orchestra/test_mcp_server.py
@@ -47,7 +47,7 @@ class TestMCPServerCreation:
 
     def test_create_mcp_server_success(self):
         """MCP server creation function exists and accepts status_service."""
-        from vibe3.orchestra.mcp_server import create_mcp_server
+        from vibe3.server.mcp import create_mcp_server
 
         # This test verifies the function exists
         # Actual MCP creation requires mcp package to be installed
@@ -66,7 +66,7 @@ class TestMCPResources:
 
     def test_status_resource_returns_json(self):
         """Status resource should return JSON snapshot."""
-        from vibe3.orchestra.mcp_server import _serialize_snapshot
+        from vibe3.server.mcp import _serialize_snapshot
 
         mock_state = MockIssueState("in-progress")
         mock_entry = MockIssueStatusEntry(
@@ -107,7 +107,7 @@ class TestMCPResources:
 
     def test_format_snapshot_for_mcp(self):
         """Snapshot should be formatted as markdown for MCP tool output."""
-        from vibe3.orchestra.mcp_server import format_snapshot_for_mcp
+        from vibe3.server.mcp import format_snapshot_for_mcp
 
         mock_state = MockIssueState("in-progress")
         mock_entry = MockIssueStatusEntry(
@@ -141,7 +141,7 @@ class TestMCPTools:
 
     def test_orchestra_status_tool(self):
         """orchestra_status tool should return formatted status."""
-        from vibe3.orchestra.mcp_server import format_snapshot_for_mcp
+        from vibe3.server.mcp import format_snapshot_for_mcp
 
         mock_state = MockIssueState("blocked")
         mock_entry = MockIssueStatusEntry(
@@ -169,7 +169,7 @@ class TestMCPTools:
     def test_orchestra_issue_detail_tool(self):
         """orchestra_issue_detail tool should return issue details."""
         # This is tested indirectly via _serialize_snapshot
-        from vibe3.orchestra.mcp_server import _serialize_snapshot
+        from vibe3.server.mcp import _serialize_snapshot
 
         mock_state = MockIssueState("review")
         mock_entry = MockIssueStatusEntry(
@@ -260,12 +260,11 @@ class TestMCPServerIntegration:
     def test_mcp_server_graceful_degradation_on_import_error(self):
         """Should continue without MCP if import fails."""
         from vibe3.orchestra.config import OrchestraConfig
-        from vibe3.orchestra.serve_utils import _build_server
+        from vibe3.server.registry import _build_server
 
         config = OrchestraConfig()
 
         # This test verifies graceful degradation
-        # The try/except in serve_utils handles ImportError gracefully
         heartbeat, fastapi_app = _build_server(config)
 
         # FastAPI app should be created even if MCP fails
@@ -276,7 +275,7 @@ class TestMCPServerIntegration:
     def test_build_server_creates_fastapi_app(self):
         """_build_server should create FastAPI app with status endpoint."""
         from vibe3.orchestra.config import OrchestraConfig
-        from vibe3.orchestra.serve_utils import _build_server
+        from vibe3.server.registry import _build_server
 
         config = OrchestraConfig()
         heartbeat, fastapi_app = _build_server(config)
