+++ tests/vibe3/orchestra/test_pr_review_dispatch.py
@@ -5,8 +5,8 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from vibe3.orchestra.config import OrchestraConfig
-from vibe3.orchestra.event_bus import GitHubEvent
 from vibe3.orchestra.services.pr_review_dispatch import PRReviewDispatchService
+from vibe3.runtime.event_bus import GitHubEvent
 
 
 class _ImmediateLoop:
