+++ tests/vibe3/orchestra/test_webhook_handler.py
@@ -8,8 +8,8 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
 from vibe3.orchestra.config import OrchestraConfig
-from vibe3.orchestra.heartbeat import HeartbeatServer
-from vibe3.orchestra.webhook_handler import _verify_signature, make_webhook_router
+from vibe3.runtime.heartbeat import HeartbeatServer
+from vibe3.server.app import _verify_signature, make_webhook_router
 
 
 def _make_app(secret: str | None = None) -> tuple[TestClient, HeartbeatServer]:
